# openapi.json
{
  "components": {
    "schemas": {
...
      "DocumentType": {
        "enum": [
          "article",
          "post",
          "thread",
          "paper",
          "pdf",
          "video",
          "page"
        ],
        "title": "DocumentType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InterestChoice": {
        "properties": {
          "level": {
            "$ref": "#/components/schemas/InterestLevel"
          },
          "topic_id": {
            "title": "Topic Id",
            "type": "integer"
          }
        },
        "required": [
          "topic_id",
          "level"
        ],
        "title": "InterestChoice",
        "type": "object"
      },
      "InterestLevel": {
        "description": "How strongly a user picked an interest in the survey (PLAN.md \u00a77 step 1); stored as the\nweight INTEREST_LEVEL_WEIGHTS gives it.",
        "enum": [
          "interested",
          "very_interested"
        ],
        "title": "InterestLevel",
        "type": "string"
      },
...
      "PinIn": {
        "properties": {
          "site": {
            "title": "Site",
            "type": "string"
          }
        },
        "required": [
          "site"
        ],
        "title": "PinIn",
        "type": "object"
      },
      "PinOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain_id": {
            "title": "Domain Id",
            "type": "integer"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "source": {
            "$ref": "#/components/schemas/PinSource"
          }
        },
        "required": [
          "domain_id",
          "host",
          "source",
          "name",
          "created_at"
        ],
        "title": "PinOut",
        "type": "object"
      },
      "PinSource": {
        "enum": [
          "survey",
          "suggested"
        ],
        "title": "PinSource",
        "type": "string"
      },
      "RankingPreset": {
        "description": "Survey step 5 (PLAN.md \u00a77); RANKING_PRESETS maps each to ranking weights.",
        "enum": [
          "balanced",
          "trust_my_sources",
          "match_my_interests",
          "fresh"
        ],
        "title": "RankingPreset",
        "type": "string"
      },
      "RankingWeights": {
        "additionalProperties": false,
        "description": "Weights of the ranking components (PLAN.md \u00a76.6); all non-negative, the hide penalty\nis subtracted.",
        "properties": {
          "feedback": {
            "minimum": 0.0,
            "title": "Feedback",
            "type": "number"
          },
          "hide": {
            "minimum": 0.0,
            "title": "Hide",
            "type": "number"
          },
          "interest": {
            "minimum": 0.0,
            "title": "Interest",
            "type": "number"
          },
          "ppr": {
            "minimum": 0.0,
            "title": "Ppr",
            "type": "number"
          },
          "recency": {
            "minimum": 0.0,
            "title": "Recency",
            "type": "number"
          }
        },
        "required": [
          "interest",
          "ppr",
          "feedback",
          "recency",
          "hide"
        ],
        "title": "RankingWeights",
        "type": "object"
      },
...
      "SurveyAnswers": {
        "additionalProperties": false,
        "description": "Version 1 of the survey, one field per step.",
        "properties": {
          "content_types": {
            "items": {
              "$ref": "#/components/schemas/DocumentType"
            },
            "minItems": 1,
            "title": "Content Types",
            "type": "array"
          },
          "exploration_pct": {
            "title": "Exploration Pct",
            "type": "number"
          },
          "interests": {
            "items": {
              "$ref": "#/components/schemas/InterestChoice"
            },
            "minItems": 1,
            "title": "Interests",
            "type": "array"
          },
          "preset": {
            "$ref": "#/components/schemas/RankingPreset"
          },
          "sites": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Sites",
            "type": "array"
          }
        },
        "required": [
          "interests",
          "content_types",
          "exploration_pct",
          "preset"
        ],
        "title": "SurveyAnswers",
        "type": "object"
      },
      "SurveyOut": {
        "properties": {
          "answers": {
            "$ref": "#/components/schemas/SurveyAnswers"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "answers",
          "created_at"
        ],
        "title": "SurveyOut",
        "type": "object"
      },
      "SurveySubmission": {
        "properties": {
          "answers": {
            "$ref": "#/components/schemas/SurveyAnswers"
          },
          "version": {
            "const": 1,
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "answers"
        ],
        "title": "SurveySubmission",
        "type": "object"
      },
...
      "UserSettingsIn": {
        "properties": {
          "content_types": {
            "items": {
              "$ref": "#/components/schemas/DocumentType"
            },
            "minItems": 1,
            "title": "Content Types",
            "type": "array"
          },
          "exploration_pct": {
            "title": "Exploration Pct",
            "type": "number"
          },
          "interests": {
            "items": {
              "$ref": "#/components/schemas/InterestChoice"
            },
            "title": "Interests",
            "type": "array"
          },
          "preset": {
            "$ref": "#/components/schemas/RankingPreset"
          },
          "summaries_opt_in": {
            "title": "Summaries Opt In",
            "type": "boolean"
          }
        },
        "required": [
          "preset",
          "exploration_pct",
          "content_types",
          "summaries_opt_in",
          "interests"
        ],
        "title": "UserSettingsIn",
        "type": "object"
      },
      "UserSettingsOut": {
        "properties": {
          "content_types": {
            "items": {
              "$ref": "#/components/schemas/DocumentType"
            },
            "title": "Content Types",
            "type": "array"
          },
          "exploration_choices": {
            "items": {
              "type": "number"
            },
            "title": "Exploration Choices",
            "type": "array"
          },
          "exploration_pct": {
            "title": "Exploration Pct",
            "type": "number"
          },
          "exploration_split_semantic": {
            "title": "Exploration Split Semantic",
            "type": "number"
          },
          "interests": {
            "items": {
              "$ref": "#/components/schemas/InterestChoice"
            },
            "title": "Interests",
            "type": "array"
          },
          "preset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RankingPreset"
              },
              {
                "type": "null"
              }
            ]
          },
          "summaries_opt_in": {
            "title": "Summaries Opt In",
            "type": "boolean"
          },
          "weights": {
            "$ref": "#/components/schemas/RankingWeights"
          }
        },
        "required": [
          "preset",
          "weights",
          "exploration_pct",
          "exploration_choices",
          "exploration_split_semantic",
          "content_types",
          "summaries_opt_in",
          "interests"
        ],
        "title": "UserSettingsOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Discovery Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/pins": {
      "get": {
        "operationId": "list_pins",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PinOut"
                  },
                  "title": "Response List Pins",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Pins",
        "tags": [
          "pins"
        ]
      },
      "post": {
        "description": "Pin a site (idempotent); its homepage and feeds join the crawl frontier.",
        "operationId": "add_pin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Pin",
        "tags": [
          "pins"
        ]
      }
    },
    "/pins/{domain_id}": {
      "delete": {
        "description": "Unpin a site. What was crawled stays in the shared graph; the user's scores follow at\nthe next cycle.",
        "operationId": "remove_pin",
        "parameters": [
          {
            "in": "path",
            "name": "domain_id",
            "required": true,
            "schema": {
              "title": "Domain Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Pin",
        "tags": [
          "pins"
        ]
      }
    },
...
    "/settings": {
      "get": {
        "description": "The user's settings, or the defaults before the survey.",
        "operationId": "user_settings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "User Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "update_user_settings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update User Settings",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/survey": {
      "post": {
        "description": "Store the answers and apply them: they replace the user's interests, pins and\nsettings (summaries stay as they were: off unless opted in on the settings page).",
        "operationId": "submit_survey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurveySubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurveyOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Survey",
        "tags": [
          "survey"
        ]
      }
    },
...
  }
}